Redefined some variables to cchange them from define to static constants, modified enemys in order to be able to create bosses, and complex enemies. Also modified particles to fit origginal shoot trace.

+++ Space Harrier/ModuleShadow.h
@@ -15,7 +15,7 @@ public:
 	bool Start();
 	bool CleanUp();
 
-	void DrawShadow(int x, int y,float scale);
+	void DrawShadow(float x, float y,float scale);
 
 private:
 	SDL_Texture* shadow;
